pyproject.toml
[project]
name = "KennyFasa_bookkeeping_software_Backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "apscheduler>=3.11.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.124.0",
    "gunicorn>=23.0.0",
    "passlib>=1.7.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "redis>=7.1.0",
    "shortuuid>=1.0.13",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.27",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]